requirement.txt
djangorestframework
psycopg2-binary
requests
python==3.7
django==2.2.5
